pyproject.toml
[project]
name = "36questions-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "firebase-admin>=6.7.0",
    "mcp[cli]>=1.5.0",
    "pydantic-ai>=0.0.43",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.27.1",
]
